requirements.txt
absl-py==2.1.0
addict==2.4.0
alabaster==0.7.16
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
bce-python-sdk==0.9.4
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
coverage==7.3.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.10.0
Cython==3.0.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
docutils==0.20.1
easyocr==1.7.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
facexlib==0.3.0
fastapi==0.112.0
fastapi-cli==0.0.5
fastjsonschema==2.19.1
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.2
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
fsspec==2024.2.0
ftfy==6.1.3
future==0.18.3
gast==0.5.4
google-api-core==2.17.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-cloud-vision==3.7.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
idna==2.10
imageio==2.34.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
ipykernel==6.29.2
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lapx==0.5.10.post1
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.43.0
lmdb==1.4.1
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.1.0
notebook_shim==0.2.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.52.2
opencv-contrib-python==4.9.0.80
opencv-python==4.10.0.84
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
paddleocr==2.6.0.1
paddlepaddle==2.6.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
premailer==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
Pygments==2.17.2
pyparsing==3.1.1
pystackreg==0.2.7
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
rarfile==4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
roboflow==1.1.19
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.37.2
style==1.1.0
supervision==0.18.0
sympy==1.13.2
termcolor==2.4.0
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing-extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.25
ultralytics-thop==2.0.0
update==0.0.1
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.30.6
uvloop==0.19.0
visualdl==2.5.3
watchfiles==0.23.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.44.0
wrapt==1.14.1
yapf==0.40.2
zipp==3.17.0
